# Series-Tracker-1.0.openapi.yaml
openapi: "3.0.3"
info:
  title: "Series Tracker API"
  version: "1.0"
  description: |-
    Series Tracker 1.0 API for finding TV series, managing a personal watchlist, and tracking watched episodes. This document describes the current backend implementation.
    
    Authentication: register through /api/register, log in through /api/login, and send Authorization: Bearer <accessToken> to protected endpoints.
    Access tokens expire after 15 minutes; refresh tokens expire after 7 days. /api/refresh returns only a new accessToken. Tokens are sent in JSON; cookie-based authentication is not implemented.
    
    Series, season, and episode IDs are separate TMDB identifiers. seasonNumber and episodeNumber are sequential numbers, not TMDB IDs. Image fields contain relative TMDB paths or null.
    
    Before adding a series to the watchlist, load /api/series/{id} to persist it. Before toggling an episode, load /api/series/{id}/season/{seasonNumber} to persist the season's episodes.
    
    Implementation details that affect clients:
    - /api/series/{id}/watched returns an array of episode IDs in watchedEpisodesCount.
    - POST /user/series returns the user's email in userSeriesId.
    - POST /api/user/episodes/{episodeId}/status toggles state and ignores the request body, including isWatched.
    - Empty searches and absent personal watchlist entries may return 200.
    - Unhandled errors, including JSON parsing failures, are returned as 500.
servers:
  -
...
    variables:
...
security: []
paths:
...
  "/api/series":
    get:
      tags: ["Series"]
      operationId: "getPopularSeries"
      summary: "Get popular TV series"
      description: |-
        Returns a plain array from the first page of TMDB /tv/popular. This endpoint does not expose pagination.
        Results are cached in memory and on disk. This request does not create Series records in the database.
        An invalid maxAge falls back to 12h instead of returning 400.
      parameters:
        -
          name: "maxAge"
          in: "query"
          required: false
          description: |-
            Maximum cache age: a nonnegative integer followed by h (hours), m (minutes), or d (days).
            A number without a suffix is interpreted as hours. Surrounding whitespace is trimmed.
            Examples: 12h, 30m, 2d, 12. Missing or invalid values default to 12h.
            0, 0h, 0m, or 0d force a fresh fetch. Suffixes are case-sensitive; 12H is not recognized.
          schema:
            type: "string"
            default: "12h"
          examples:
            hours:
              value: "12h"
            minutes:
              value: "30m"
            days:
              value: "2d"
            forceRefresh:
              value: "0"
      responses:
        "200":
          description: "Popular series loaded from cache or TMDB."
          content:
            "application/json":
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PopularSeries"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          description: "Failed to fetch popular series from TMDB."
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PopularSeriesError"
              example:
                error: "Failed to fetch series from TMDb"
                details: "TMDb request failed: 503 Service Unavailable"
  "/api/series/{id}":
    get:
      tags: ["Series"]
      operationId: "getSeriesDetails"
      summary: "Get series details and season summaries"
      description: |-
        Reads the series from the database first. If it is missing, fetches it from TMDB and stores it with seasons whose seasonNumber is greater than 0.
        Existing database records are returned without refreshing their TMDB metadata. Seasons are sorted by seasonNumber in ascending order. Episodes are not loaded by this endpoint.
        There is no explicit validation of id; an invalid value may result in 500.
      parameters:
        -
          $ref: "#/components/parameters/SeriesId"
      responses:
        "200":
          description: "Series details."
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/SeriesDetails"
        "404":
          description: "The series was not found in TMDB."
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Series not found"
        "409":
          $ref: "#/components/responses/DatabaseConflict"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/UpstreamError"
  "/api/series/{id}/season/{seasonNumber}":
    get:
      tags: ["Series"]
      operationId: "getSeasonEpisodes"
      summary: "Get a season and its episodes"
      description: |-
        Returns the stored season when it already contains episodes. Otherwise, loads the series if needed, fetches the season from TMDB, and stores its episodes.
        Episodes are sorted by their sequential number in ascending order. A stored nonempty season is not refreshed on subsequent requests.
        The response uses episodeNumber. It does not include episodeNum or isWatched. Read personal viewing state through /api/series/{id}/watched.
        Numeric path parameters have no explicit validation. Season 0 is excluded during the initial series import, so special episodes are not reliably supported by this endpoint.
      parameters:
        -
          $ref: "#/components/parameters/SeriesId"
        -
          name: "seasonNumber"
          in: "path"
          required: true
          description: "Sequential regular-season number, starting at 1. This is not a TMDB season ID."
          schema:
            type: "integer"
          example: 1
      responses:
        "200":
          description: "Season metadata and episodes."
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/SeasonWithEpisodes"
        "404":
          description: "The series or season was not found, or a related database record is missing."
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                seriesMissing:
                  value:
                    error: "Series not found"
                seasonMissing:
                  value:
                    error: "Seasons not found"
                relationMissing:
                  value:
                    error: "User or Series not found in database"
        "409":
          $ref: "#/components/responses/DatabaseConflict"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/UpstreamError"
  "/api/series/{id}/watched":
    get:
      tags: ["Series"]
      operationId: "getWatchedEpisodeIds"
      summary: "Get watched episode IDs for a series"
      description: |-
        Returns TMDB IDs of episodes watched by the current user.
        Despite its name, watchedEpisodesCount is an array of numbers, not a count.
        Returns an empty array when no matching watch records exist. The series itself is not separately checked for existence. ID ordering is unspecified.
        There is no explicit validation of id.
      security:
        -
          bearerAuth: []
      parameters:
        -
          $ref: "#/components/parameters/SeriesId"
      responses:
        "200":
          description: "Watched episode IDs."
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/WatchedEpisodeIdsResponse"
              examples:
                watched:
                  value:
                    watchedEpisodesCount: [1001,1002]
                empty:
                  value:
                    watchedEpisodesCount: []
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: |-
        Use the access token issued by /api/login or /api/refresh.
        Send Authorization: Bearer <accessToken>. The Bearer prefix is case-sensitive in the current middleware.
        In Swagger UI, enter only the token in the Authorize dialog. Access tokens expire after 15 minutes.
  parameters:
    SeriesId:
      name: "id"
      in: "path"
      required: true
      description: "TMDB series ID. /api/series endpoints do not explicitly validate this path parameter."
      schema:
        type: "integer"
      example: 1396
...
  responses:
...
    Unauthorized:
      description: |-
        Access denied: the Authorization header is missing, the prefix is incorrect, or the token is empty.
        Invalid token: the JWT is invalid or expired.
        For a verified token without a userId claim, individual controllers may return 'Unauthorized' or 'Unauthorized access'.
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                error: "Access denied"
            invalid:
              value:
                error: "Invalid token"
...
    DatabaseConflict:
      description: "A database uniqueness constraint was violated. Usually caused by adding the same series twice; it can also occur during concurrent operations. The shared error handler determines this message."
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "This serial is already in your list"
    InternalError:
      description: "An unexpected server or database error occurred."
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Internal server error"
    UpstreamError:
      description: |-
        A TMDB error with the HTTP status returned by the upstream service.
        Series details use 'Series not found'; season details use 'Seasons not found'. Search forwards TMDB status_message or a fallback error message.
        Possible upstream responses include 400, 401, 403, 422, 429, and 5xx.
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: "object"
      required: ["error"]
      properties:
        error:
          type: "string"
          description: "Human-readable error message. The current API does not return a separate machine-readable error code."
      example:
        error: "Internal server error"
...
    SeriesStatus:
      type: "string"
      description: "Production status. TMDB statuses containing 'ended' map to ended; statuses containing 'cancel' map to canceled; all other values map to ongoing."
      enum: ["ongoing","ended","canceled"]
...
    ImagePath:
      type: "string"
      description: "Relative TMDB image path, or null when no image is available. Add the TMDB image-service base URL to display the image."
      nullable: true
      example: "/poster.jpg"
    NullableOverview:
      type: "string"
      description: "Series, season, or episode overview. May be null."
      nullable: true
      example: "A short plot summary."
    NullableDateTime:
      type: "string"
      description: "ISO 8601 UTC date-time, or null when the date is unavailable."
      format: "date-time"
      nullable: true
      example: "2008-01-20T00:00:00.000Z"
    PopularSeries:
      type: "object"
      required: ["id","title","poster","backdrop","description","rating","releaseDate"]
      properties:
        id:
          type: "integer"
          description: "TMDB series ID. This endpoint uses id rather than tmdbId."
          example: 1396
        title:
          type: "string"
          description: "Series title."
          example: "Breaking Bad"
        poster:
          $ref: "#/components/schemas/ImagePath"
        backdrop:
          $ref: "#/components/schemas/ImagePath"
        description:
          type: "string"
          description: "TMDB overview. This endpoint calls the field description rather than overview."
        rating:
          type: "number"
          description: "The vote_average value returned by TMDB."
          example: 8.8
        releaseDate:
          type: "string"
          description: "TMDB first_air_date, usually YYYY-MM-DD. May be an empty string."
          example: "2008-01-20"
    PopularSeriesError:
      type: "object"
      required: ["error","details"]
      properties:
        error:
          type: "string"
          enum: ["Failed to fetch series from TMDb"]
        details:
          type: "string"
          description: "Underlying fetch error message."
    SeasonSummary:
      type: "object"
      required: ["tmdbId","seasonNumber","name","overview","poster"]
      properties:
        tmdbId:
          type: "integer"
          description: "TMDB season ID. This is different from its sequential season number."
          example: 3572
        seasonNumber:
          type: "integer"
          description: "Sequential season number."
          example: 1
        name:
          type: "string"
          description: "Season name."
          example: "Season 1"
        overview:
          $ref: "#/components/schemas/NullableOverview"
        poster:
          $ref: "#/components/schemas/ImagePath"
    SeriesDetails:
      type: "object"
      required: ["tmdbId","title","poster","backdrop","overview","firstAirDate","numberOfSeasons","numberOfEpisodes","status","seasons"]
      properties:
        tmdbId:
          type: "integer"
          description: "TMDB series ID."
          example: 1396
        title:
          type: "string"
          description: "Series title."
          example: "Breaking Bad"
        poster:
          $ref: "#/components/schemas/ImagePath"
        backdrop:
          $ref: "#/components/schemas/ImagePath"
        overview:
          $ref: "#/components/schemas/NullableOverview"
        firstAirDate:
          $ref: "#/components/schemas/NullableDateTime"
        numberOfSeasons:
          type: "integer"
          description: "Total season count from stored TMDB metadata."
          nullable: true
          example: 5
        numberOfEpisodes:
          type: "integer"
          description: "Total episode count from stored TMDB metadata."
          nullable: true
          example: 62
        status:
          $ref: "#/components/schemas/SeriesStatus"
        seasons:
          type: "array"
          description: "Stored seasons sorted by seasonNumber in ascending order."
          items:
            $ref: "#/components/schemas/SeasonSummary"
    Episode:
      type: "object"
      required: ["tmdbId","seasonNumber","episodeNumber","title","overview","stillPath","airDate"]
      properties:
        tmdbId:
          type: "integer"
          description: "TMDB episode ID used by the watch-status toggle endpoint."
          example: 1001
        seasonNumber:
          type: "integer"
          description: "Sequential season number."
          example: 1
        episodeNumber:
          type: "integer"
          description: "Sequential episode number within the season. This is the actual field name in this response."
          example: 1
        title:
          type: "string"
          description: "Episode title."
          example: "Pilot"
        overview:
          $ref: "#/components/schemas/NullableOverview"
        stillPath:
          $ref: "#/components/schemas/ImagePath"
        airDate:
          $ref: "#/components/schemas/NullableDateTime"
    SeasonWithEpisodes:
      allOf:
        -
          $ref: "#/components/schemas/SeasonSummary"
        -
          type: "object"
          required: ["episodes"]
          properties:
            episodes:
              type: "array"
              description: "Episodes sorted by episodeNumber in ascending order. Personal isWatched values are not included."
              items:
                $ref: "#/components/schemas/Episode"
    WatchedEpisodeIdsResponse:
      type: "object"
      required: ["watchedEpisodesCount"]
      properties:
        watchedEpisodesCount:
          type: "array"
          description: "Despite the field name, this is an array of watched TMDB episode IDs, not a numeric count."
          items:
            type: "integer"
      example:
        watchedEpisodesCount: [1001,1002]
...